requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
sqlalchemy==2.0.40
psycopg[binary]==3.2.7
httpx==0.28.1
python-dotenv==1.0.1
jinja2==3.1.6
python-multipart==0.0.20
pydantic==2.11.3
